Rename dwelltime_analysis to dwelltime_extraction
